requirements.txt
joblib==1.4.2
numpy==2.0.1
pyspellchecker==0.8.1
scikit-learn==1.5.1
scipy==1.13.1
threadpoolctl==3.5.0
